requirements.txt
pytest>=7.0.0
requests>=2.28.0
allure-pytest>=2.9.0
pytest-html>=3.1.0
python-dotenv>=0.19.0
faker>=15.0.0
selenium>=4.10.0
webdriver-manager>=4.0.0